+++ src/App.tsx
@@ -1,11 +1,12 @@
 import '@mantine/core/styles.css';
 import { MantineProvider } from '@mantine/core';
-import { Router } from './Router';
-import { theme } from './theme';
+
+import cqTheme from './theme';
+import Router from './router';
 
 export default function App() {
   return (
-    <MantineProvider theme={theme} defaultColorScheme="dark">
+    <MantineProvider theme={cqTheme} defaultColorScheme="dark">
       <Router />
     </MantineProvider>
   );
