--- src/Router.tsx
@@ -1,13 +1,0 @@
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import { HomePage } from './pages/Home.page';
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <HomePage />,
-  },
-]);
-
-export function Router() {
-  return <RouterProvider router={router} />;
-}
