+++ src/layouts/FormLayout/FormLayout.scss
@@ -1,5 +1,6 @@
 :root {
-  --form-with: 26.1875rem;
+  --form-with: 26.25rem;
+  --form-height: 20rem;
 }
 
 .cq-form-wrapper-shadow {
@@ -8,9 +9,17 @@
   width: calc(var(--form-with) + 2rem);
   position: absolute;
   z-index: 1;
-  min-height: 24.3125rem;
   margin-left: 3.5rem;
   padding: var(--mantine-spacing-xl) var(--mantine-spacing-xl);
+
+  &.admin{
+    min-height: calc(var(--form-height) + 4rem);
+  }
+
+  &.client{
+    --form-height: 13.875rem;
+    min-height: calc(var(--form-height) + 2rem);
+  }
 }
 
 .cq-form-bg {
@@ -19,6 +28,13 @@
   border-radius: var(--mantine-radius-md);
   width: var(--form-with);
   padding: var(--mantine-spacing-xl) var(--mantine-spacing-xl);
+
+  &__title {
+    display: block;
+    text-align: center;
+    margin-bottom: var(--mantine-spacing-sm);
+  }
+
   form {
     display: flex;
     flex-direction: column;
